skills: Re-sync on every source-backed skill, list and detail page
The use case is 'I just edited my skill in Drive — confirm it's in
Stash', which applies to any source-backed skill, not only drafts. The
button moves out of the draft CTA into a shared component and now sits
beside Run on runnable rows/cards, and on the skill's own page next to
'Edit it in Google Drive'. read_source_skill carries source_id so the
detail page can trigger the sync.
